requirements.txt
elasticsearch==8.10.0
requests
beautifulsoup4
kafka-python